Move command runner classes
