+++ config/sync/core.extension.yml
@@ -195,7 +195,6 @@ module:
   workbench_access: 0
   workbench_menu_access: 0
   workflows: 0
-  zippylib: 0
   limited_field_widgets: 1
   menu_link_content: 1
   pathauto: 1
