+++ docroot/modules/custom/va_gov_build_trigger/src/Command/CommandRunner.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\va_gov_content_export\ExportCommand;
+namespace Drupal\va_gov_build_trigger\Command;
 
 use Symfony\Component\Process\Process;
